Move to a New + Register interface
In order to facilitate discovery of services in applications that don't need or
want to register an endpoint, we separate out the gsr.Start() function into a
gsr.NewRegistry() and gsr.Registry.Register() method.

While doing so and updating the documentation in the README.md, I noticed that
it would be useful to have Service and Endpoint objects and have the
gsr.Registry.Endpoints() function return a list of those objects.