fix incorrect return of etcd endpoints
The etcdEndpoints() function, defined in conf.go, was mistakenly using append()
to add a new element to the back of the "res" variable. The res variable is a
[]string that had already been created with a length equal to the number of
endpoints found in the GSR_ETCD_ENDPOINTS environment variable. When
canonicalizing the endpoint values, we needed to use res[x] = ep instead of
append() because the length of res was already correctly set. The bug
manifested itself by returning a []string with the first element containing a
NULL string, which would cause etcd.clientv3.New() to try accessing an empty
endpoint which would return a dial tcp: missing address failure.