switch to dep
`govendor` was great, but the community has coalesced around `dep` so
it's time we switched over to it. This patch removes the
vendor/vendor.json `govendor` dependency tracking file and adds the
Gopkg.lock|toml files that `dep` uses, and adds better
test/fmtcheck/vet targets to the Makefile.

Close Issue #6