Add `gsr.Registry.Unregister()` method
In order to fast-remove a service endpoint from gsr, this patch adds a
`gsr.Registry.Unregister()` method that can be used in a signal trap
function to remove a service endpoint from the registry.

The patch includes code added to the example data and web applications
that shows how to handle SIGTERM and adds a `docker stop` command to the
examples/run.sh along with some text to show that the web container
should receive a notification about the data service being deleted when
the SIGTERM is received.

Issue #17